Attempt to wrap circuitpython stuff

+++ build.sh
@@ -2,8 +2,8 @@
 
 source common.sh
 
-MICRO_PYTHON_VERSION=${MICRO_PYTHON_VERSION:-v1.19.1}
-ESP_IDF_VERSION="v4.4.1"
+MICRO_PYTHON_VERSION=${MICRO_PYTHON_VERSION:-master}
+ESP_IDF_VERSION="v4.4.2"
 if [ ! -d "${build_dir}" ]; then
   mkdir "${build_dir}"
 fi
@@ -44,25 +44,34 @@ git checkout "${MICRO_PYTHON_VERSION}"
 if [ ! -f "${build_dir}/microPython/project/micropython/mpy-cross/mpy-cross" ]; then
   pushd mpy-cross
   make
-  cp mpy-cross "${venv_dir}/bin/"
+  if [ -f mpy-cross ]; then
+    cp mpy-cross "${venv_dir}/bin/"
+  elif [ -f build/mpy-cross ]; then
+    cp build/mpy-cross "${venv_dir}/bin/"
+  else
+    echo "Could not find mpy cross?"
+    exit 1
+  fi
   popd
 fi
 pushd ./ports/unix
-if [ ! -f "micorpython" ]; then
-  make submodules &> submod
-  make &> base || (cat base; exit 1)
-  export PATH="${PATH}:$(pwd)"
+if ! command -v micropython &> /dev/null; then
+  if [ ! -f "micorpython" ]; then
+    make submodules &> submod
+    make &> base || (cat base; exit 1)
+  fi
+  export PATH="${PATH}:$(pwd):$(pwd)/build-standard/"
 fi
 popd
 # Make uasyncio available for testing on unix port
 if [ ! -d ~/.micropython/lib/ ]; then
   mkdir -p ~/.micropython/lib
   cp -af ./extmod/* ~/.micropython/lib/
-  micropython -m upip install unittest logging threading typing warnings base64 hmac
+  micropython -m upip install unittest logging threading typing warnings base64 hmac  
 fi
 # Run some smoke tests
 pushd "${FW_DIR}"
-flake8 --max-line-length 100 --ignore=Q000 --ignore=W504 --exclude=manifest.py
+# flake8 --max-line-length 100 --ignore=Q000 --ignore=W504 --exclude=manifest.py
 micropython -c "import unittest;unittest.main('smoke_test')"
 popd
 pushd "${MP_ROOT}/ports/esp32"
@@ -75,6 +84,7 @@ make submodules &> submod
 make BOARD=${BOARD:-MOTOC3} FROZEN_MANIFEST="${SCRIPT_DIR}/fw/manifest.py" clean
 make USER_C_MODULES="${SCRIPT_DIR}/modules/micropython.cmake" clean
 make BOARD=${BOARD:-MOTOC3} FROZEN_MANIFEST="${SCRIPT_DIR}/fw/manifest.py"
+# make BOARD=${BOARD:-MOTOC3}
 # USER_C_MODULES="${SCRIPT_DIR}/modules/micropython.cmake"
 pwd
 popd
